Added a UUID newtype to get around the aeson/UUID requirement that UUIDs have hyphens. Watson for some reason stores them without hyphens, so for compatibility the workaround was required.
With the new type, there is custom parsing to insert hyphens and then
run the standad fromText of UUID. The newtype is required because
there's no other way that I can see to override the FromJSON instance of
UUID that's included with aeson.

Result of all this is that watson's frames file is now parsable, and
running a report gives the same results as watson.

Next: tags and change -y to be from start of year, not 365 days
   ... and lots of testing of report options, including current frame
   in report.